requirements.txt
setuptools
tensorboard
pandas
matplotlib
ruamel.yaml
gym
moviepy
ffmpeg-python
internetarchive
ipdb
lpips
einops
protobuf
dm_control
scipy
easydict
memory_maze
crafter
opencv-python
pyyaml
jinja2
lxml
coloredlogs
xmltodict
Pyro4
psutil
multiprocess
pytest
daemoniker
tqdm
requests
mypy-extensions
jsonlines
praw
wget
importlib_resources
hydra-core
Pillow
DI-toolkit
DI-treetensor
flask
urlobject
tabulate
tensorboardX
# minerl==0.4.4
# This was needed for minerl
# conda install -c conda-forge openjdk=8
